Update README.md
The KNeighborsClassifier, found in the Scikit-Learn library, is a tool for classification tasks like determining the species of Iris flowers. In Iris flower classification, you use a dataset with measurements of sepal length, sepal width, petal length, and petal width as features. The goal is to train a model to predict the species (setosa, versicolor, or virginica) based on these features. The KNeighborsClassifier relies on the k-nearest neighbors algorithm, assigning a class to a new data point based on the majority class of its nearest neighbors. This method simplifies the complex task of classifying Iris flowers by associating their unique features with specific species labels.